% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condor.R
\name{apsimOut2Nc}
\alias{apsimOut2Nc}
\title{Post-processing APSIM outputs into netcdf files}
\usage{
apsimOut2Nc(
  filename,
  years,
  factors,
  traits,
  par_factor,
  base,
  factor_files,
  base_sub = NULL,
  merge_out_nc = TRUE,
  yearvar = "year",
  cpu_num = 14,
  tmp_folder = file.path("Results", "tmpNc")
)
}
\arguments{
\item{filename}{The filename of netcdf file}

\item{years}{The years in the APSIM and netcdf}

\item{factors}{The factors in the netcdf and APSIM}

\item{traits}{The vector of traits write into netcdf file}

\item{par_factor}{The factors will be parallel processed}

\item{base}{The base folder to store APSIM output files}

\item{factor_files}{The data.frame or vector of files to specify the output file names and levels}

\item{base_sub}{The sub-folder to store the single output files}

\item{merge_out_nc}{Whether merge the parallel outputs into an array when updating netcdf files.}

\item{yearvar}{Variable name of yeas in the APSIM output file}

\item{cpu_num}{The cpu number of parallel}

\item{tmp_folder}{The tmp folder to store merge APSIM output files}
}
\description{
Post-processing APSIM outputs into netcdf files
}
