% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condor.R
\name{clusterRunPreparingTask}
\alias{clusterRunPreparingTask}
\title{Preparing task table for APSIM simulations}
\usage{
clusterRunPreparingTask(
  project,
  factors,
  sharedrive,
  local,
  par_factor,
  input_factor,
  skip_factor = NULL,
  folder_prefix = NULL,
  r_script = "R\\\\bin\\\\i386\\\\Rscript.exe",
  r_code = "RCode.R",
  ids = 0
)
}
\arguments{
\item{project}{The project name in the ClusterRun}

\item{factors}{The factors list to generate simulations}

\item{sharedrive}{sharedrive}

\item{local}{local}

\item{par_factor}{The parallel factors to write into database of ClusterRun}

\item{input_factor}{The factors to indicate input files}

\item{skip_factor}{The skip factors to be processed in the condor clients}

\item{folder_prefix}{Output folder to store the out files of all simulations}

\item{r_script}{The path to Rscript.exe file}

\item{r_code}{r_code}

\item{ids}{The starting id of simulations}
}
\description{
Preparing task table for APSIM simulations
}
