% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condor.R
\name{compressInputs}
\alias{compressInputs}
\title{Compress the input files}
\usage{
compressInputs(
  input,
  output,
  sevenz = "C:\\\\Program Files\\\\7-Zip\\\\7z.exe"
)
}
\arguments{
\item{input}{list of files and folders}

\item{output}{The file of output}

\item{sevenz}{The path to 7z.exe}
}
\description{
Compress the input files
}
