% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimumAPSIM.R
\name{minimumAPSIM}
\alias{minimumAPSIM}
\title{Create the minimum requirements to run an APSIM simulation (APSIM 7.5)}
\usage{
minimumAPSIM(template, apsim, output = "APSIM75")
}
\arguments{
\item{template}{The sim file as the template}

\item{apsim}{The installed path of APSIM}

\item{output}{The output folder of all files}
}
\description{
Create the minimum requirements to run an APSIM simulation (APSIM 7.5)
}
