% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimumR.R
\name{minimumR}
\alias{minimumR}
\title{Prepare minimum R through only copying necessary files.}
\usage{
minimumR(pkgs, output = "MR", r.home = R.home(), arch = "i386")
}
\arguments{
\item{pkgs}{A character vector of export packages}

\item{output}{The output folder of all files}

\item{r.home}{Home directory of R}

\item{arch}{The architecture of R will be copied, i368 or x64. NO implement now.}
}
\description{
Prepare minimum R through only copying necessary files.
}
