% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncAPSIM.R
\name{ncClose}
\alias{ncClose}
\title{Synonymy with nc_close in ncdf4 package}
\usage{
ncClose(nc)
}
\arguments{
\item{nc}{An object of class ncdf4 (as returned from nc_open), indicating what file to read from.}
}
\description{
Synonymy with nc_close in ncdf4 package
}
