% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncAPSIM.R
\name{ncCreateApsim}
\alias{ncCreateApsim}
\title{Create netcdf file for APSIM}
\usage{
ncCreateApsim(
  filename,
  years,
  factors,
  traits,
  is.close = TRUE,
  global.att = NULL,
  ...
)
}
\arguments{
\item{filename}{The filename to save netcdf file.}

\item{years}{A vector for year demension.}

\item{factors}{A list for all dimensions in the netcdf.}

\item{traits}{A vector of traits in the netcdf.}

\item{is.close}{Whether to close netcdf file.}

\item{global.att}{Global attribures}

\item{...}{Attributes of traits, e.g. unit
If false, a nv variable will be returned.}
}
\description{
Create netcdf file for APSIM
}
