% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncAPSIM.R
\name{ncFactorLevels}
\alias{ncFactorLevels}
\title{Obtain factor levels from a netCDF file}
\usage{
ncFactorLevels(nc, factorname = NA, split = " *, *")
}
\arguments{
\item{nc}{An object of class ncdf4 (as returned from nc_open), indicating what file to read from.}

\item{factorname}{The leves of factor name will be returned. All levels will be returned if NA is specified.}

\item{split}{character called by strsplit}
}
\description{
Obtain factor levels from a netCDF file
}
