% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncAPSIM.R
\name{ncGetDimNames}
\alias{ncGetDimNames}
\title{Get all dimension names for one variables}
\usage{
ncGetDimNames(nc, var)
}
\arguments{
\item{nc}{An object of class ncdf4 (as returned from nc_open), indicating what file to read from.}
}
\description{
Get all dimension names for one variables
}
