% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncAPSIM.R
\name{ncGetStart}
\alias{ncGetStart}
\title{Obtain start according dimension levels}
\usage{
ncGetStart(nc, ...)
}
\arguments{
\item{nc}{nc}

\item{...}{extra arguments}
}
\description{
Obtain start according dimension levels
}
