% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncAPSIM.R
\name{ncGetTraits}
\alias{ncGetTraits}
\title{Obtain trait value from a netCDF file}
\usage{
ncGetTraits(nc, traits = NULL, ...)
}
\arguments{
\item{nc}{An object of class ncdf4 (as returned from nc_open), indicating what file to read from.}

\item{traits}{The trait name will be returned}

\item{...}{Other arguments for factors. 
All variables will be returned in this factor if a factor is not specified.}
}
\description{
Obtain trait value from a netCDF file
}
