% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncAPSIM.R
\name{ncOpen}
\alias{ncOpen}
\title{Synonymy with nc_open in ncdf4 package}
\usage{
ncOpen(filename, ...)
}
\arguments{
\item{filename}{Name of the existing netCDF file to be opened.}

\item{...}{Other arguments pass to nc_open}
}
\description{
Synonymy with nc_open in ncdf4 package
}
