% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncAPSIM.R
\name{ncReadApsimIndex}
\alias{ncReadApsimIndex}
\title{Convert factor levels to index}
\usage{
ncReadApsimIndex(nc, filename, title = "Title", split = ",", ...)
}
\arguments{
\item{nc}{nc}

\item{filename}{filename}

\item{title}{title}

\item{split}{split}

\item{...}{...}
}
\description{
Convert factor levels to index
}
