% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncAPSIM.R
\name{ncReadApsimOut}
\alias{ncReadApsimOut}
\title{Read APSIM out file}
\usage{
ncReadApsimOut(
  nc,
  filenames,
  start,
  res.array,
  yearvar = "Year",
  traits = NULL,
  ...
)
}
\arguments{
\item{nc}{nc}

\item{filenames}{filenames}

\item{start}{start}

\item{res.array}{res.array}

\item{yearvar}{yearvar}

\item{traits}{traits}

\item{...}{extra arguments}
}
\description{
Read APSIM out file
}
