% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncAPSIM.R
\name{ncSubset}
\alias{ncSubset}
\title{Obtain a subset from several nc files.}
\usage{
ncSubset(ncs, filename, traits = NULL, ...)
}
\arguments{
\item{ncs}{A vector of nc files}

\item{filename}{Filename for new nc file}

\item{traits}{A vector of traits will be retured. All traits will be return if NULL}

\item{...}{Dimensions will be returned}
}
\description{
Obtain a subset from several nc files.
}
