% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncAPSIM.R
\name{ncTraitList}
\alias{ncTraitList}
\title{Obtain trait list from a netCDF file}
\usage{
ncTraitList(nc, attname = "TraitList", split = " *, *")
}
\arguments{
\item{nc}{An object of class ncdf4 (as returned from nc_open), indicating what file to read from.}

\item{attname}{The default attribute name of trait list}

\item{split}{character called by strsplit}
}
\description{
Obtain trait list from a netCDF file
}
